+++ userService/src/main/java/com/wefit/userService/exception/UserConflictException.java
@@ -1,0 +1,7 @@
+package com.wefit.userService.exception;
+
+public class UserConflictException extends RuntimeException {
+    public UserConflictException(String message) {
+        super(message);
+    }
+}